+++ backend/app.js
@@ -2,10 +2,12 @@ require("dotenv").config();
 const express = require("express");
 const path = require("path");
 // const server = require('./server')
+const cors = require("cors");
+const app = express();
+app.use(cors());
 
 const { sequelize } = require("./models");
 
-const app = express();
 app.set("port", process.env.PORT || 3000);
 
 sequelize
@@ -20,7 +22,7 @@ app.use(express.static(path.join(__dirname, "public")));
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
-const Router = require("./routes/router");
-app.use("/api", Router);
+const Router = require("./routes");
+app.use("/api/v1", Router);
 
 module.exports = app;
